Move Pokedex render into Vue
