+++ .editorconfig
@@ -11,7 +11,7 @@ trim_trailing_whitespace = true
 
 # Matches multiple files with brace expansion notation
 # Set default charset and indentation style
-[*.{ts,tsx,js,html,ejs,css,scss,less}]
+[*.{ts,tsx,js,html,ejs,vue,css,scss,less}]
 charset = utf-8
 indent_style = space
 indent_size = 4
