+++ package.json
@@ -42,6 +42,7 @@
     "terser-webpack-plugin": "^5.1.1",
     "vue": "^3.0.7",
     "vue-loader": "^16.1.2",
+    "vuex": "^4.0.0",
     "webpack": "^5.24.2",
     "webpack-merge": "^5.7.3",
     "webpack-nano": "^1.1.1",
