+++ src/index.js
@@ -23,7 +23,7 @@ document.addEventListener('DOMContentLoaded', () => {
 
 // load everything we need
 const lastSave = Date.now();
-const player = Player(lastSave);
+const player = Player(lastSave, models.app);
 const { Poke, makeRandomPoke } = mkPoke(player);
 const enemy = makeEnemy(null, player, Poke);
 const combatLoop = Combat(player, enemy);
